client/asset/btc: move spv scan filters code to sep file
